refactor(go-server): split extension inventories by extension
