+++ packages/server/duckdb-server-go/pkg/functionset/avro.go
@@ -1,0 +1,8 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-avro/tree/f9d590297485f0318f480372c70bdd852826e258
+var avroComputeFunctions = [...]string{}
+
+var avroElevatedFunctions = [...]string{
+	"read_avro",
+}
