+++ packages/server/duckdb-server-go/pkg/functionset/aws.go
@@ -1,0 +1,8 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-aws/tree/efa54a990e16c976576685dd4134d2478cf5a574
+var awsComputeFunctions = [...]string{}
+
+var awsElevatedFunctions = [...]string{
+	"load_aws_credentials",
+}
