+++ packages/server/duckdb-server-go/pkg/functionset/azure.go
@@ -1,0 +1,6 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-azure/tree/003214c96d0caa39d5c3e27a9e1976a0692c7d37
+var azureComputeFunctions = [...]string{}
+
+var azureElevatedFunctions = [...]string{}
