+++ packages/server/duckdb-server-go/pkg/functionset/encodings.go
@@ -1,0 +1,6 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-encodings/tree/06295e77b13de65842992c82f14289ea679e4730
+var encodingsComputeFunctions = [...]string{}
+
+var encodingsElevatedFunctions = [...]string{}
