+++ packages/server/duckdb-server-go/pkg/functionset/excel.go
@@ -1,0 +1,11 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-excel/tree/f4c72b5ef04a03b3a78a95b5a2ee94ba93e3178d
+var excelComputeFunctions = [...]string{
+	"excel_text",
+	"text",
+}
+
+var excelElevatedFunctions = [...]string{
+	"read_xlsx",
+}
