+++ packages/server/duckdb-server-go/pkg/functionset/fts.go
@@ -1,0 +1,11 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-fts/tree/6814ec9a7d5fd63500176507262b0dbf7cea0095
+var ftsComputeFunctions = [...]string{
+	"stem",
+}
+
+var ftsElevatedFunctions = [...]string{
+	"create_fts_index",
+	"drop_fts_index",
+}
