+++ packages/server/duckdb-server-go/pkg/functionset/httpfs.go
@@ -1,0 +1,6 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-httpfs/tree/827222fb45a043a7a852d1f7aae46901492a3cda
+var httpfsComputeFunctions = [...]string{}
+
+var httpfsElevatedFunctions = [...]string{}
