+++ packages/server/duckdb-server-go/pkg/functionset/inet.go
@@ -1,0 +1,18 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-inet/tree/fe7f60bb60245197680fb07ecd1629a1dc3d91c8
+var inetComputeFunctions = [...]string{
+	"+",
+	"-",
+	"<<=",
+	">>=",
+	"broadcast",
+	"family",
+	"host",
+	"html_escape",
+	"html_unescape",
+	"netmask",
+	"network",
+}
+
+var inetElevatedFunctions = [...]string{}
