+++ packages/server/duckdb-server-go/pkg/functionset/json.go
@@ -1,16 +1,6 @@
 package functionset
 
 // These inventories match DuckDB 1.5.5 at d8cdaa33fda8df955cc76ef58a280f68f4cd43fa.
-var autocompleteComputeFunctions = [...]string{
-	"check_peg_parser",
-}
-
-var autocompleteElevatedFunctions = [...]string{
-	"disable_peg_parser",
-	"enable_peg_parser",
-	"sql_auto_complete",
-}
-
 var jsonComputeFunctions = [...]string{
 	"->>",
 	"array_to_json",
@@ -57,24 +47,4 @@ var jsonElevatedFunctions = [...]string{
 	"read_ndjson",
 	"read_ndjson_auto",
 	"read_ndjson_objects",
-}
-
-var tpcdsComputeFunctions = [...]string{
-	"tpcds_answers",
-	"tpcds_queries",
-}
-
-var tpcdsElevatedFunctions = [...]string{
-	"dsdgen",
-	"tpcds",
-}
-
-var tpchComputeFunctions = [...]string{
-	"tpch_answers",
-	"tpch_queries",
-}
-
-var tpchElevatedFunctions = [...]string{
-	"dbgen",
-	"tpch",
 }
