+++ packages/server/duckdb-server-go/pkg/functionset/mysql.go
@@ -1,0 +1,12 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-mysql/tree/7267164dab3409e943261aeee6ae32f1b00847a7
+var mysqlComputeFunctions = [...]string{}
+
+var mysqlElevatedFunctions = [...]string{
+	"mysql_clear_cache",
+	"mysql_debug_execution_plan",
+	"mysql_execute",
+	"mysql_explain_federated",
+	"mysql_query",
+}
