+++ packages/server/duckdb-server-go/pkg/functionset/sqlite.go
@@ -1,0 +1,10 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-sqlite/tree/f79b1db7d7730b18d0f8400d3650ffa6b45168d8
+var sqliteComputeFunctions = [...]string{}
+
+var sqliteElevatedFunctions = [...]string{
+	"sqlite_attach",
+	"sqlite_query",
+	"sqlite_scan",
+}
