+++ packages/server/duckdb-server-go/pkg/functionset/tpch.go
@@ -1,0 +1,12 @@
+package functionset
+
+// These inventories match DuckDB 1.5.5 at d8cdaa33fda8df955cc76ef58a280f68f4cd43fa.
+var tpchComputeFunctions = [...]string{
+	"tpch_answers",
+	"tpch_queries",
+}
+
+var tpchElevatedFunctions = [...]string{
+	"dbgen",
+	"tpch",
+}
