+++ packages/server/duckdb-server-go/pkg/functionset/ui.go
@@ -1,0 +1,11 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-ui/tree/26084e6e02cc7f47ba0361daef646d75974908d9
+var uiComputeFunctions = [...]string{}
+
+var uiElevatedFunctions = [...]string{
+	"start_ui",
+	"start_ui_server",
+	"stop_ui_server",
+	"ui_is_started",
+}
