+++ packages/server/duckdb-server-go/pkg/functionset/unity_catalog.go
@@ -1,0 +1,11 @@
+package functionset
+
+// Source: https://github.com/duckdb/unity_catalog/tree/fd851475780ca064d9706a5025ea6e5d1d9d7e23
+var unityCatalogComputeFunctions = [...]string{}
+
+var unityCatalogElevatedFunctions = [...]string{
+	"__internal_delta_ccv2_commit_staged",
+	"table_data_path",
+	"unity_catalog_checkpoint_table",
+	"unity_catalog_force_checkpoint_table",
+}
