+++ packages/server/duckdb-server-go/pkg/functionset/vss.go
@@ -1,0 +1,12 @@
+package functionset
+
+// Source: https://github.com/duckdb/duckdb-vss/tree/b833341c8737fd3f3558c7720cc575ae8fc82598
+var vssComputeFunctions = [...]string{}
+
+var vssElevatedFunctions = [...]string{
+	"hnsw_compact_index",
+	"hnsw_index_scan",
+	"pragma_hnsw_index_info",
+	"vss_join",
+	"vss_match",
+}
